LLM-agent/requirements.txt
# LangChain Auto-GPT 에이전트 필수 라이브러리
# Core LangChain packages
langchain>=0.1.0
langchain-experimental>=0.0.50
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector database and embeddings
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# API integrations
openai>=1.0.0
google-search-results>=2.4.2

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development tools (optional)
streamlit>=1.28.0
